pyproject.toml
[tool.poetry]
name = "precommit_imagesize"
version = "0.1.0"
description = ""
authors = ["alex <alex@pdp7.net>"]
license = "Public Domain"

[tool.poetry.dependencies]
python = "^3.6"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
check-image-size = 'precommit_imagesize:main'
